requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.8.0
certifi==2024.7.4
charset-normalizer==2.1.1
click==8.1.7
cloudscraper==1.2.71
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
executing==2.0.1
fastcore==1.6.3
fastlite==0.0.7
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
gotrue==2.8.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
instructor>=1.6.0
ipykernel==6.29.5
ipython==8.26.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kaleido==0.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.0.1
oauthlib==3.2.2
openai==1.40.2
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
plotly==5.23.0
postgrest==0.16.11
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.2.4
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
realtime==2.0.2
referencing==0.35.1
regex==2024.7.24
replicate==0.31.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.20.0
setuptools==69.0.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
sqlite-vec==0.1.6
sqlite_minutils==3.36.0.post4
stack-data==0.6.3
starlette==0.38.2
storage3==0.7.7
StrEnum==0.4.15
supabase==2.8.0
supafunc==0.5.1
tenacity==8.5.0
tiktoken==0.7.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
wcwidth==0.2.13
websockets==12.0
wheel==0.44.0
yarl==1.9.4
psycopg2-binary==2.9.10
anthropic==0.39.0
google-generativeai==0.8.3
groq==0.11.0
